@@ -2,7 +2,7 @@
 #
 # The Pyrocko Developers, 21st Century
 # ---|P------/S----------~Lg----------
-from __future__ import absolute_import
+
 from pyrocko.util import parse_md
 
 __doc__ = parse_md(__file__)
