@@ -25,8 +25,6 @@ selection by the user. This list currently contains the following sites:
 Any other site can be specified by providing its full URL.
 '''
 
-from __future__ import absolute_import
-
 import re
 import logging
 import socket
@@ -38,11 +36,6 @@ from pyrocko import config
 
 from pyrocko.util import \
     urlencode, Request, build_opener, HTTPDigestAuthHandler, urlopen, HTTPError
-
-try:
-    newstr = unicode
-except NameError:
-    newstr = str
 
 
 logger = logging.getLogger('pyrocko.client.fdsn')
@@ -240,7 +233,7 @@ def _request(
 
     req = Request(url)
     if post:
-        if isinstance(post, newstr):
+        if isinstance(post, str):
             post = post.encode('utf8')
         logger.debug('POST data: \n%s' % post.decode('utf8'))
         req.data = post
@@ -650,7 +643,7 @@ def event(
             raise InvalidRequest(
                 'If parsed=True is selected, format="xml" must be selected.')
 
-        return quakeml.load_xml(stream=fh)
+        return quakeml.QuakeML.load_xml(stream=fh)
 
     else:
         return fh
