@@ -2,6 +2,6 @@
 #
 # The Pyrocko Developers, 21st Century
 # ---|P------/S----------~Lg----------
-from __future__ import absolute_import
+
 from .report_main import *  # noqa
 from .report_call import *  # noqa
