@@ -2,7 +2,7 @@
 #
 # The Pyrocko Developers, 21st Century
 # ---|P------/S----------~Lg----------
-from __future__ import absolute_import
+
 from ..snuffling import Snuffling, Param, Switch
 import numpy as num
 from matplotlib import cm
