@@ -5,7 +5,7 @@
 '''
 File IO module for SICK traces format.
 '''
-from __future__ import division, absolute_import
+
 import os
 from struct import unpack
 
