@@ -1,4 +1,7 @@
-from __future__ import absolute_import, print_function
+# http://pyrocko.org - GPLv3
+#
+# The Pyrocko Developers, 21st Century
+# ---|P------/S----------~Lg----------
 
 import sys
 import time
