@@ -3,11 +3,8 @@
 # The Pyrocko Developers, 21st Century
 # ---|P------/S----------~Lg----------
 
-from __future__ import absolute_import, print_function
-
-
 from . import base, selection, database, model, io, client, tool, error, \
-    environment, dataset, operators
+    environment, dataset, operators, check
 
 from .base import *  # noqa
 from .selection import *  # noqa
@@ -20,8 +17,9 @@ from .error import *  # noqa
 from .environment import *  # noqa
 from .dataset import *  # noqa
 from .operators import *  # noqa
+from .check import *  # noqa
 
 __all__ = base.__all__ + selection.__all__ + database.__all__ \
     + model.__all__ + io.__all__ + client.__all__ + tool.__all__ \
     + error.__all__ + environment.__all__ + dataset.__all__ \
-    + operators.__all__
+    + operators.__all__ + check.__all__
