@@ -3,8 +3,6 @@
 # The Pyrocko Developers, 21st Century
 # ---|P------/S----------~Lg----------
 
-from __future__ import absolute_import, print_function
-
 import logging
 import math
 import os
@@ -1617,6 +1615,34 @@ class Squirrel(Selection):
             sources=[s.describe() for s in self._sources],
             operators=[op.describe() for op in self._operators])
 
+    @filldocs
+    def check(
+            self, obj=None, tmin=None, tmax=None, time=None, codes=None,
+            ignore=[]):
+        '''
+        Check for common data/metadata problems.
+
+        %(query_args)s
+
+        :param ignore:
+            Problem types to be ignored.
+        :type ignore:
+            :class:`list` of :class:`str`
+            (:py:class:`~pyrocko.squirrel.check.SquirrelCheckProblemType`)
+
+        :returns:
+            :py:class:`~pyrocko.squirrel.check.SquirrelCheck` object
+            containing the results of the check.
+
+        See :py:func:`~pyrocko.squirrel.check.do_check`.
+        '''
+
+        from .check import do_check
+        tmin, tmax, codes = self._get_selection_args(
+            CHANNEL, obj, tmin, tmax, time, codes)
+
+        return do_check(self, tmin=tmin, tmax=tmax, codes=codes, ignore=ignore)
+
     def get_content(
             self,
             nut,
@@ -2505,10 +2531,16 @@ class Squirrel(Selection):
 
         from pyrocko import model as pmodel
 
+        if codes is not None:
+            codes = codes_patterns_for_kind(STATION, codes)
+
         by_nsl = defaultdict(lambda: (list(), list()))
         for station in self.get_stations(obj, tmin, tmax, time, codes):
             sargs = station._get_pyrocko_station_args()
             by_nsl[station.codes.nsl][0].append(sargs)
+
+        if codes is not None:
+            codes = [model.CodesNSLCE(c) for c in codes]
 
         for channel in self.get_channels(obj, tmin, tmax, time, codes):
             sargs = channel._get_pyrocko_station_args()
