@@ -3,11 +3,8 @@
 # The Pyrocko Developers, 21st Century
 # ---|P------/S----------~Lg----------
 
-from __future__ import absolute_import, print_function
-
 import time
 import logging
-from builtins import str as newstr
 
 from pyrocko import util
 from pyrocko.io.io_common import FileLoadError
@@ -204,7 +201,7 @@ def iload(
     selection = None
     kind_ids = to_kind_ids(content)
 
-    if isinstance(paths, (str, newstr)):
+    if isinstance(paths, str):
         paths = [paths]
     else:
         if segment is not None:
