@@ -42,9 +42,8 @@ def _weed(dists, badnesses, neighborhood=1, interaction_radius=3.,
     if ndeleted == 0:
         return deleted
 
-    kept = num.logical_not(deleted).nonzero()[0]
-
-    xdists = dists[num.meshgrid(kept, kept)]
+    kept = num.logical_not(deleted)
+    xdists = dists[kept][:, kept]
     xbadnesses = badnesses[kept]
     xdeleted = _weed(xdists, xbadnesses, neighborhood, interaction_radius,
                      del_frac, max_del-ndeleted, max_depth, depth+1)
@@ -73,8 +72,8 @@ def weed(x, y, badnesses, neighborhood=1, nwanted=None, interaction_radius=3.):
         dists, badnesses, neighborhood, interaction_radius, del_frac=4,
         max_del=n-nwanted, max_depth=500, depth=0)
 
-    kept = num.logical_not(deleted).nonzero()[0]
-    dists_kept = dists[num.meshgrid(kept, kept)]
+    kept = num.logical_not(deleted)
+    dists_kept = dists[kept][:, kept]
     meandists_kept = neighborhood_density(dists_kept, neighborhood)
     return deleted, meandists_kept
 
