@@ -1,4 +1,3 @@
-from __future__ import division, print_function, absolute_import
 
 import unittest
 import logging
@@ -50,7 +49,7 @@ mantle
         self.tempdirs.append(store_dir)
 
         qsconf = qseis.QSeisConfig()
-        qsconf.qseis_version = '2006a'
+        qsconf.qseis_version = '2006b'
 
         qsconf.time_region = (
             gf.meta.Timing('0'),
@@ -74,7 +73,7 @@ mantle
             distance_min=550*km,
             distance_max=560*km,
             distance_delta=1*km,
-            modelling_code_id='qseis.2006a',
+            modelling_code_id='qseis.2006b',
             earthmodel_1d=mod,
             tabulated_phases=[
                 gf.meta.TPDef(
