@@ -1,4 +1,3 @@
-from __future__ import division, print_function, absolute_import
 import unittest
 import math
 import tempfile
@@ -156,8 +155,8 @@ class ModelTestCase(unittest.TestCase):
 
             r = g(projected, 'R')
             t = g(projected, 'T')
-            assert(near(r.ydata[0], 1.0, 0.001))
-            assert(near(t.ydata[0], 1.0, 0.001))
+            assert near(r.ydata[0], 1.0, 0.001)
+            assert near(t.ydata[0], 1.0, 0.001)
 
     def testProjectionsZOnly(self):
         km = 1000.
@@ -188,7 +187,7 @@ class ModelTestCase(unittest.TestCase):
                         return tr
 
             z = g(projected, 'U')
-            assert(near(z.ydata[0], 1.0, 0.001))
+            assert near(z.ydata[0], 1.0, 0.001)
 
     def getGNSSCampaign(self, nstations=25, nsparse=False):
         rstate = num.random.RandomState(None)
@@ -290,7 +289,7 @@ class ModelTestCase(unittest.TestCase):
 
             idx += ncomp
 
-        num.testing.assert_array_equal(corr_arr, corr_arr_ref)
+        num.testing.assert_allclose(corr_arr, corr_arr_ref, rtol=1e-5)
 
     def testGNSSCampaignCovarianceMatrix(self):
         campaign = self.getGNSSCampaign(2)
@@ -324,7 +323,7 @@ class ModelTestCase(unittest.TestCase):
 
             idx += ncomp
 
-        num.testing.assert_array_equal(covar_arr, covar_ref)
+        num.testing.assert_allclose(covar_arr, covar_ref, rtol=1e-5)
 
     def testGNSSCampaignSparse(self):
         nstations = 20
